add AuthenticationWrapper
